requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
celery-types==0.23.0
cffi==1.17.1
channels==4.2.2
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.4
Django==5.0.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
msgpack==1.1.1
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2025.7.34
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.2.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zope.interface==7.2
